pps: try harder to get pps-gpio irq
Call platform_get_irq() if gpiod_to_irq() fails. This is intended for
use with SoCs, like the amlogic s905x, that don't implement
gpiod_to_irq but do support a gpiochip interrupt controller driver.

This is a working alternate pps-gpio dts config that works on GPIOAO_9
on the AML-S905X-CC board:

    pps-gpio {
        compatible = "pps-gpio";
        gpios = <&gpio_ao GPIOAO_9 GPIO_ACTIVE_HIGH>;
        interrupt-parent = <&gpio_intc>;
        interrupts = <9 1>;
        assert-rising-edge;
        status = "okay";
    };